Separated the inner classes and attempted to work on SUBMITNAME
