+++ BBCA/day5_bca/ChatServer.java
@@ -1,4 +1,4 @@
-package day5_bca;
+package BBCA;
 
 import java.net.Inet4Address;
 import java.net.ServerSocket;
@@ -14,7 +14,7 @@ import java.io.PrintWriter;
 
 public class ChatServer {
     public static final int PORT = 54321;
-    private static final ArrayList<ClientConnectionData> clientList = new ArrayList<>();
+    public static final ArrayList<ClientConnectionData> clientList = new ArrayList<>();
 
     public static void main(String[] args) throws Exception {
         ExecutorService pool = Executors.newFixedThreadPool(100);
@@ -44,7 +44,7 @@ public class ChatServer {
                     System.out.println("added client " + name);
 
                     //handle client business in another thread
-                    pool.execute(new ClientHandler(client));
+                    pool.execute(new ServerClientHandler(client));
                 } catch (IOException ex) {
                     System.out.println(ex.getMessage());
                 }
@@ -53,80 +53,5 @@ public class ChatServer {
         } 
     }
 
-    // Inner class 
-    static class ClientHandler implements Runnable {
-        // Maintain data about the client serviced by this thread
-        ClientConnectionData client;
-
-        public ClientHandler(ClientConnectionData client) {
-            this.client = client;
-        }
-
-        /**
-		 * Broadcasts a message to all clients connected to the server.
-		 */
-        public void broadcast(String msg) {
-            try {
-                System.out.println("Broadcasting -- " + msg);
-                synchronized (clientList) {
-                    for (ClientConnectionData c : clientList){
-                        c.getOut().println(msg);
-                        // c.getOut().flush();
-                    }
-                }
-            } catch (Exception ex) {
-                System.out.println("broadcast caught exception: " + ex);
-                ex.printStackTrace();
-            }
-            
-        }
-
-        @Override
-        public void run() {
-            try {
-                BufferedReader in = client.getInput();
-                //get userName, first message from user
-                String userName = in.readLine().trim();
-                client.setUserName(userName);
-                //notify all that client has joined
-                broadcast(String.format("WELCOME %s", client.getUserName()));
-
-                
-                String incoming = "";
-
-                while( (incoming = in.readLine()) != null) {
-                    if (incoming.startsWith("CHAT")) {
-                        String chat = incoming.substring(4).trim();
-                        if (chat.length() > 0) {
-                            String msg = String.format("CHAT %s %s", client.getUserName(), chat);
-                            broadcast(msg);    
-                        }
-                    } else if (incoming.startsWith("QUIT")){
-                        break;
-                    }
-                }
-            } catch (Exception ex) {
-                if (ex instanceof SocketException) {
-                    System.out.println("Caught socket ex for " + 
-                        client.getName());
-                } else {
-                    System.out.println(ex);
-                    ex.printStackTrace();
-                }
-            } finally {
-                //Remove client from clientList, notify all
-                synchronized (clientList) {
-                    clientList.remove(client); 
-                }
-                System.out.println(client.getName() + " has left.");
-                broadcast(String.format("EXIT %s", client.getUserName()));
-                try {
-                    client.getSocket().close();
-                } catch (IOException ex) {}
-
-            }
-        }
-        
-    }
 
 }
