+++ BBCA/day5_bca/ServerClientHandler.java
@@ -1,0 +1,96 @@
+package BBCA;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.net.SocketException;
+
+import static BBCA.ChatServer.clientList;
+
+
+public class ServerClientHandler implements Runnable{
+
+    // Maintain data about the client serviced by this thread
+    ClientConnectionData client;
+
+    public ServerClientHandler(ClientConnectionData client) {
+        this.client = client;
+    }
+
+    /**
+     * Broadcasts a message to all clients connected to the server.
+     */
+    public void broadcast(String msg) {
+        try {
+            System.out.println("Broadcasting -- " + msg);
+            synchronized (clientList) {
+                for (ClientConnectionData c : clientList){
+                    c.getOut().println(msg);
+                    // c.getOut().flush();
+                }
+            }
+        } catch (Exception ex) {
+            System.out.println("broadcast caught exception: " + ex);
+            ex.printStackTrace();
+        }
+
+    }
+
+    @Override
+    public void run() {
+        try {
+            BufferedReader in = client.getInput();
+            String userName;
+
+            while (!(userName = in.readLine()).startsWith("NAME") || userName.substring(5).trim().equals("hello")){
+                broadcast("SUBMITNAME");
+            }
+
+            //get userName, first message from user
+            userName = userName.substring(5).trim();
+
+            System.out.println(userName);
+
+            client.setUserName(userName);
+            //notify all that client has joined
+            broadcast(String.format("WELCOME %s", client.getUserName()));
+
+
+            String incoming = "";
+
+            while( (incoming = in.readLine()) != null) {
+                System.out.println(incoming);
+
+
+                if (incoming.startsWith("CHAT")) {
+                    String chat = incoming.substring(4).trim();
+                    if (chat.length() > 0) {
+                        String msg = String.format("CHAT %s %s", client.getUserName(), chat);
+                        broadcast(msg);
+                    }
+                } else if (incoming.startsWith("QUIT")){
+                    break;
+                }
+            }
+        } catch (Exception ex) {
+            if (ex instanceof SocketException) {
+                System.out.println("Caught socket ex for " +
+                        client.getName());
+            } else {
+                System.out.println(ex);
+                ex.printStackTrace();
+            }
+        } finally {
+            //Remove client from clientList, notify all
+            synchronized (clientList) {
+                clientList.remove(client);
+            }
+            System.out.println(client.getName() + " has left.");
+            broadcast(String.format("EXIT %s", client.getUserName()));
+            try {
+                client.getSocket().close();
+            } catch (IOException ex) {}
+
+        }
+    }
+
+}
